requirements.txt
allure-pytest == 2.13.2
selene == 2.0.0rc4
pytest == 7.4.2